pyproject.toml
[tool.poetry]
name = "asistant-python"
version = "0.1.0"
description = ""
authors = ["ronihdzz <ronaldo.runing_@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
websockets = "^13.1"
twilio = "^9.3.3"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
